Extract duplicate test helpers to shared scope
