+++ EssentialFeedTests/Feed API/URLSessionHTTPClientTests.swift
@@ -137,16 +137,9 @@ class URLSessionHTTPClientTests: XCTestCase {
         return receivedResult
     }
     
-    private func anyURL() -> URL {
-        return URL(string: "http://any-url.com")!
-    }
     
     private func anyData() -> Data {
         Data(bytes: "any data".utf8)
-    }
-    
-    private func anyNSError() -> NSError {
-        NSError(domain: "any error", code: 0)
     }
     
     private func anyHTTPURLResponse() -> HTTPURLResponse {
