+++ EssentialFeedTests/Helpers/FeedCacheTestHelpers.swift
@@ -1,0 +1,30 @@
+//
+//  FeedCacheTestHelpers.swift
+//  EssentialFeedTests
+//
+//  Created by Macbook Pro on 6/9/23.
+//
+
+import Foundation
+import EssentialFeed
+
+func uniqueImage() -> FeedImage {
+    return FeedImage(id: UUID(), description: "any", location: "any", url: anyURL())
+}
+
+func uniqueImageFeed() -> (models: [FeedImage], local: [LocalFeedImage]) {
+    let items = [uniqueImage(), uniqueImage()]
+    let localItems = items.map { LocalFeedImage(id: $0.id, description: $0.description, location: $0.location, url: $0.url)}
+    
+    return (items, localItems)
+}
+
+extension Date {
+    func adding(days: Int) -> Date {
+        return Calendar(identifier: .gregorian).date(byAdding: .day, value: days, to: self)!
+    }
+    
+    func adding(seconds: TimeInterval) -> Date {
+        return Calendar(identifier: .gregorian).date(byAdding: .second, value: Int(seconds), to: self)!
+    }
+}
